no/edgebox_final/requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.7.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-celery==3.3.1
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-redis==4.10.0
django-timezone-field==3.0
djangorestframework==3.9.4
eventlet==0.23.0
flower==0.9.3
greenlet==0.4.15
idna==2.8
importlib-metadata==0.19
itsdangerous==1.1.0
kombu==3.0.37
modbus-tk==1.1.0
netifaces==0.10.9
numpy==1.17.3
paho-mqtt==1.4.0
psutil==5.6.3
Pygments==2.4.2
PyMySQL==0.9.3
pyserial==3.4
python-crontab==2.3.7
python-dateutil==2.8.0
pytz==2019.1
redis==2.10.6
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tornado==5.1.1
urllib3==1.25.3
vine==1.3.0
WMI==1.4.9
zipp==0.5.2
